packages/hooke-tx/pyproject.toml
[project]
name = "hooke-tx"
version = "0.1.0"
description = "Transcriptomics perturbation prediction"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
  # Core: torch + geometric
  "torch>=2.0",
  "torchvision",
  "torchaudio",
  "torch-geometric",
  "torch-scatter",
  "torch-sparse",
  "torch-cluster",
  "torch-spline-conv",
  "pyg-lib",
  # Training & experiment
  "lightning>=2.0",
  "hydra-core>=1.3",
  "hydra-submitit-launcher",
  "wandb",
  "torchmetrics",
  "torch-optimizer",
  # Scientific stack
  "numpy",
  "scipy",
  "pandas",
  "polars",
  "matplotlib",
  "seaborn",
  "scikit-learn",
  "h5py",
  "zarr>=3",
  "einops",
  "tqdm",
  "pyyaml",
  "toml",
  "python-dotenv",
  "rich",
  "joblib",
  "omegaconf",
  "networkx",
  # Flow matching
  "torchdiffeq",
  "torch-ema",
  "geomloss",
  "diffusers",
  # Chemistry
  "rdkit",
  "datamol",
  "molfeat",
  "mordredcommunity",
  "selfies",
  "safe-mol",
  # Transcriptomics
  "scanpy",
  "anndata",
  # Viz / utils
  "plotly",
  "umap-learn",
  "e3nn",
  "ipykernel",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "ruff>=0.11",
  "mypy>=1.0",
  "pre-commit>=4",
]

[tool.uv]
find-links = [
  "https://data.pyg.org/whl/torch-2.6.0+cu124.html",
  "https://data.pyg.org/whl/torch-2.6.0+cpu.html",
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
